fix eager python c code gen
