+++ paddle/phi/ops/yaml/python_api_info.yaml
@@ -8,10 +8,10 @@
   args_alias :
     use_default_mapping : True
 
-# - op : matmul
-#   name : [paddle.matmul,paddle.Tensor.matmul]
-#   args_alias :
-#     use_default_mapping : True
+- op : matmul
+  name : [paddle.matmul,paddle.Tensor.matmul]
+  args_alias :
+    use_default_mapping : True
 - op : multiply
   name : [paddle.multiply,paddle.Tensor.multiply]
   args_alias :
